+++ backend/bang-ggood/src/main/java/com/bang_ggood/auth/controller/AuthController.java
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class AuthController {
 
-
     private static final Logger log = LoggerFactory.getLogger(AuthController.class);
     private final AuthService authService;
     private final CookieProvider cookieProvider;
