+++ backend/bang-ggood/src/main/java/com/bang_ggood/auth/service/AuthService.java
@@ -2,17 +2,9 @@ package com.bang_ggood.auth.service;
 
 import com.bang_ggood.auth.dto.request.OauthLoginRequest;
 import com.bang_ggood.auth.dto.response.OauthInfoApiResponse;
-import com.bang_ggood.checklist.dto.request.ChecklistRequest;
-import com.bang_ggood.checklist.service.ChecklistManageService;
+import com.bang_ggood.global.DefaultChecklistService;
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
-import com.bang_ggood.option.domain.Option;
-import com.bang_ggood.question.domain.Answer;
-import com.bang_ggood.question.domain.CustomChecklistQuestion;
-import com.bang_ggood.question.domain.Question;
-import com.bang_ggood.question.dto.request.QuestionRequest;
-import com.bang_ggood.question.repository.CustomChecklistQuestionRepository;
-import com.bang_ggood.room.dto.request.RoomRequest;
 import com.bang_ggood.user.domain.User;
 import com.bang_ggood.user.domain.UserType;
 import com.bang_ggood.user.repository.UserRepository;
@@ -21,7 +13,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Service
@@ -32,25 +23,15 @@ public class AuthService {
 
     private final OauthClient oauthClient;
     private final JwtTokenProvider jwtTokenProvider;
-    private final ChecklistManageService checklistManageService;
+    private final DefaultChecklistService defaultChecklistService;
     private final UserRepository userRepository;
-    private final CustomChecklistQuestionRepository customChecklistQuestionRepository;
 
     public AuthService(OauthClient oauthClient, JwtTokenProvider jwtTokenProvider,
-                       ChecklistManageService checklistManageService,
-                       UserRepository userRepository,
-                       CustomChecklistQuestionRepository customChecklistQuestionRepository) {
+                       DefaultChecklistService defaultChecklistService, UserRepository userRepository) {
         this.oauthClient = oauthClient;
         this.jwtTokenProvider = jwtTokenProvider;
-        this.checklistManageService = checklistManageService;
+        this.defaultChecklistService = defaultChecklistService;
         this.userRepository = userRepository;
-        this.customChecklistQuestionRepository = customChecklistQuestionRepository;
-    }
-
-    private static void validateTokenOwnership(User user, AuthUser authUser) {
-        if (!user.getId().equals(authUser.id())) {
-            throw new BangggoodException(ExceptionCode.AUTHENTICATION_TOKEN_NOT_OWNED_BY_USER);
-        }
     }
 
     @Transactional
@@ -65,45 +46,9 @@ public class AuthService {
 
     private User signUp(OauthInfoApiResponse oauthInfoApiResponse) {
         User user = userRepository.save(oauthInfoApiResponse.toUserEntity());
-        createDefaultChecklistQuestions(user);
-        createDefaultChecklist(user);
+        defaultChecklistService.createDefaultChecklistQuestions(user);
+        defaultChecklistService.createDefaultChecklist(user);
         return user;
-    }
-
-    private void createDefaultChecklistQuestions(User user) { //TODO 리팩토링 필요
-        List<CustomChecklistQuestion> checklistQuestions = Question.findDefaultQuestions()
-                .stream()
-                .map(question -> new CustomChecklistQuestion(user, question))
-                .toList();
-
-        customChecklistQuestionRepository.saveAll(checklistQuestions);
-    }
-
-    private void createDefaultChecklist(User user) {
-        RoomRequest roomRequest = new RoomRequest(
-                "예시용 체크리스트", "방끗시 집잘구하구 행복하동", "방방하우스", "잠실", 10,
-                3000, 60, 5, List.of(1, 3), "지상", 14, "분리형 원룸", 9.5,
-                12, 9, "초", "방끗공인중개사",
-                "이곳에 필요한 메모를 작성하세요.", "이곳에 한줄평을 남겨 보세요.");
-
-        List<Integer> options = List.of(
-                Option.INDUCTION.getId(),
-                Option.AIR_CONDITIONER.getId(),
-                Option.SINK.getId(),
-                Option.BED.getId());
-
-        List<QuestionRequest> questionRequests = List.of(
-                new QuestionRequest(Question.ROOM_CONDITION_1.getId(), Answer.GOOD.name()),
-                new QuestionRequest(Question.ROOM_CONDITION_2.getId(), Answer.BAD.name()),
-                new QuestionRequest(Question.ROOM_CONDITION_3.getId(), Answer.GOOD.name()),
-                new QuestionRequest(Question.WINDOW_1.getId(), Answer.GOOD.name()),
-                new QuestionRequest(Question.WINDOW_2.getId(), Answer.BAD.name()),
-                new QuestionRequest(Question.BATHROOM_1.getId(), Answer.GOOD.name()),
-                new QuestionRequest(Question.BATHROOM_2.getId(), Answer.GOOD.name()));
-
-        ChecklistRequest checklistRequest = new ChecklistRequest(roomRequest, options, questionRequests);
-        checklistManageService.createChecklist(user, checklistRequest);
-        //TODO: 로직 리팩토링 필요
     }
 
     @Transactional(readOnly = true)
@@ -120,10 +65,17 @@ public class AuthService {
         blackList.add(splitToken);
     }
 
+    private static void validateTokenOwnership(User user, AuthUser authUser) {
+        if (!user.getId().equals(authUser.id())) {
+            throw new BangggoodException(ExceptionCode.AUTHENTICATION_TOKEN_NOT_OWNED_BY_USER);
+        }
+    }
+
     public boolean isAccessTokenInBlackList(String accessToken) {
         return blackList.contains(accessToken);
     }
 
+    @Transactional(readOnly = true)
     public User getAuthUser(String token) {
         log.info("extractUser token: {}", token);
         AuthUser authUser = jwtTokenProvider.resolveToken(token);
